dip_minor_project/requirements.txt
streamlit
opencv-python-headless
pywavelets
pillow
numpy
